wrap all wifi operations in control panel with listening mode
